Split up function runtime context integration tests into multiple files.
