+++ tests/integration/functions/runtime_context/string_parameters_test.ts
@@ -1,0 +1,156 @@
+import {
+  assert, IsExact
+} from "../../../../src/dev_deps.ts";
+import { CanBe, CanBeUndefined } from "../../../../src/test_utils.ts";
+import { DefineFunction } from "../../../../src/mod.ts";
+import {
+  EnrichedSlackFunctionHandler,
+} from "../../../../src/functions/types.ts";
+import { SlackFunctionTester } from "../../../../src/functions/tester/mod.ts";
+import { assertEqualsTypedValues } from "../../../../src/test_utils.ts";
+
+/**
+ * Custom Function handler tests, exercising string-based inputs and outputs
+ */
+Deno.test("Custom Function with a string input and no outputs should provide the string input in a function handler context", () => {
+  const TestFn = DefineFunction({
+    callback_id: "test",
+    title: "test fn",
+    source_file: "test.ts",
+    input_parameters: {
+      properties: {
+        in: {
+          type: "string",
+        },
+      },
+      required: ["in"],
+    },
+  });
+  const handler: EnrichedSlackFunctionHandler<typeof TestFn.definition> = (
+    { inputs },
+  ) => {
+    assert<IsExact<typeof inputs.in, string>>(true);
+    return {
+      outputs: {},
+    };
+  };
+  const { createContext } = SlackFunctionTester(TestFn);
+  const inputs = { in: "test" };
+  const result = handler(createContext({ inputs }));
+  assertEqualsTypedValues(result.outputs, {});
+});
+
+Deno.test("Custom Function with a string input and output should provide the string input in a function handler context", () => {
+  const TestFn = DefineFunction({
+    callback_id: "test",
+    title: "test fn",
+    source_file: "test.ts",
+    input_parameters: {
+      properties: {
+        in: {
+          type: "string",
+        },
+      },
+      required: ["in"],
+    },
+    output_parameters: {
+      properties: {
+        out: {
+          type: "string",
+        },
+      },
+      required: ["out"],
+    },
+  });
+  const handler: EnrichedSlackFunctionHandler<typeof TestFn.definition> = (
+    { inputs },
+  ) => {
+    assert<IsExact<typeof inputs.in, string>>(true);
+    return {
+      outputs: {
+        out: inputs.in,
+      },
+    };
+  };
+  const { createContext } = SlackFunctionTester(TestFn);
+  const inputs = { in: "test" };
+  const result = handler(createContext({ inputs }));
+
+  const stringTest = SlackFunctionTester("test");
+  const stringResult = handler(stringTest.createContext({ inputs }));
+
+  assertEqualsTypedValues(
+    result.outputs?.out,
+    inputs.in,
+  );
+  assertEqualsTypedValues(
+    inputs.in,
+    stringResult.outputs?.out,
+  );
+});
+
+Deno.test("Custom Function with an optional string input and output should provide the string/undefined input in a function handler context", () => {
+  const TestFn = DefineFunction({
+    callback_id: "test",
+    title: "test fn",
+    source_file: "test.ts",
+    input_parameters: {
+      properties: {
+        in: {
+          type: "string",
+        },
+      },
+      required: [],
+    },
+    output_parameters: {
+      properties: {
+        out: {
+          type: "string",
+        },
+      },
+      required: ["out"],
+    },
+  });
+  const handler: EnrichedSlackFunctionHandler<typeof TestFn.definition> = (
+    { inputs },
+  ) => {
+    assert<CanBe<typeof inputs.in, string>>(true);
+    assert<CanBeUndefined<typeof inputs.in>>(true);
+    return {
+      outputs: {
+        out: inputs.in || "default",
+      },
+    };
+  };
+  const { createContext } = SlackFunctionTester(TestFn);
+  const inputs = {};
+  const result = handler(createContext({ inputs }));
+  assertEqualsTypedValues(result.outputs?.out, "default");
+});
+
+Deno.test("Custom Function with only string output", () => {
+  const TestFn = DefineFunction({
+    callback_id: "test",
+    title: "test fn",
+    source_file: "test.ts",
+    output_parameters: {
+      properties: {
+        out: {
+          type: "string",
+        },
+      },
+      required: ["out"],
+    },
+  });
+  const handler: EnrichedSlackFunctionHandler<typeof TestFn.definition> =
+    () => {
+      return {
+        outputs: {
+          out: "test",
+        },
+      };
+    };
+  const { createContext } = SlackFunctionTester(TestFn);
+  const result = handler(createContext({ inputs: {} }));
+  assertEqualsTypedValues(result.outputs?.out, "test");
+});
